+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The public API of this library consists of the functions declared in file
 [h3api.h.in](./src/h3lib/include/h3api.h.in).
 
 ## [Unreleased]
+### Added
+- Added `getIndexDigit` function for inspecting specific index digits of an H3 index. (#1024)
 
 ## [4.3.0] - 2025-06-17
 ### Added